+++ CHANGELOG.md
@@ -11,6 +11,8 @@
   * New metric type: Object ([#2796](https://github.com/mozilla/glean/pull/2796))
 * iOS
   * New metric type: Object ([#2796](https://github.com/mozilla/glean/pull/2796))
+* Python
+  * New metric type: Object ([#2796](https://github.com/mozilla/glean/pull/2796))
 
 # v59.0.0 (2024-03-28)